requirements.txt
requests
flask
nudenet
Pillow
mcrcon
openai
qq-botpy